Fixed code smells after running sonarqube on the project
Multiple code smells were found by SonarQube on the project. Some of them were violations of the below rules:

Local variable names in java must be with the default regular expression ^[a-z][a-zA-Z0-9]*$ RSPEC-117
The diamond operator ("<>") should be used when the type of the maps are already defined RSPEC-2293
Package names being camel cased is a code smell, all package names must be lowercase
String literals should not be duplicated RSPEC-1192
Generic types shouldn't be used raw (without type parameters) in variable declarations or return values. Doing so bypasses generic type checking, and defers the catch of unsafe code to runtime. RSPEC-3740
The dangling else problem appears when nested if/else statements are written without curly braces. In this case, else is associated with the nearest if but that is not always obvious and sometimes the indentation can also be misleading.RSPEC-5261
Here is the overall SonarQube report for the repo (Development) branch post the fixes.